+++ modules/ismp/pallets/relayer/src/lib.rs
@@ -695,9 +695,10 @@ where
                             StateMachine::Kusama(_) |
                             StateMachine::Polkadot(_) => {
                                 use codec::Decode;
-                                let receipt =
-                                    pallet_ismp::ResponseReceipt::decode(&mut &*encoded_receipt)
-                                        .map_err(|_| Error::<T>::ProofValidationError)?;
+                                let receipt = pallet_ismp::primitives::ResponseReceipt::decode(
+                                    &mut &*encoded_receipt,
+                                )
+                                .map_err(|_| Error::<T>::ProofValidationError)?;
                                 (receipt.relayer, receipt.response.0)
                             },
                         }
