+++ modules/ismp/testsuite/Cargo.toml
@@ -6,7 +6,7 @@ description = "Test suite for  ISMP"
 authors = ["Polytope Labs <hello@polytope.technology>"]
 
 [dependencies]
-ismp = { path = "../core" }
+ismp = {  workspace = true }
 primitive-types = "0.12.1"
 codec = { package = "parity-scale-codec", version = "3.1.3" }
 sp-core = { workspace = true, default-features = true }
