+++ parachain/runtimes/gargantua/src/lib.rs
@@ -88,7 +88,7 @@ use weights::{BlockExecutionWeight, ExtrinsicBaseWeight, RocksDbWeight};
 use ::staging_xcm::latest::prelude::BodyId;
 use cumulus_primitives_core::ParaId;
 use frame_support::{derive_impl, traits::ConstBool};
-use pallet_ismp::{mmr::Leaf, ProofKeys};
+use pallet_ismp::mmr::{Leaf, ProofKeys};
 
 /// Alias to 512-bit hash when used in the context of a transaction signature on the chain.
 pub type Signature = MultiSignature;
